+++ core/src/main/java/edu/jhu/hlt/concrete/summarization/SummarySourceType.java
@@ -4,24 +4,31 @@
  * DO NOT EDIT UNLESS YOU ARE SURE THAT YOU KNOW WHAT YOU ARE DOING
  *  @generated
  */
-package edu.jhu.hlt.concrete.services;
+package edu.jhu.hlt.concrete.summarization;
 
 
 import java.util.Map;
 import java.util.HashMap;
 import org.apache.thrift.TEnum;
 
-/**
- * Annotation Tasks Types
- */
-public enum AnnotationTaskType implements org.apache.thrift.TEnum {
-  TRANSLATION(1),
-  NER(2),
-  TOPICID(3);
+public enum SummarySourceType implements org.apache.thrift.TEnum {
+  /**
+   * Specifies that sourceIds is a list of Communication.UUIDs.
+   * This can be used for single or multi-document summarization.
+   */
+  DOCUMENT(0),
+  /**
+   * Specifies that sourceIds is a list of Tokenization.UUIDs.
+   */
+  TOKENIZATION(1),
+  /**
+   * Specifies that sourceIds is a list of Entity.UUIDs
+   */
+  ENTITY(2);
 
   private final int value;
 
-  private AnnotationTaskType(int value) {
+  private SummarySourceType(int value) {
     this.value = value;
   }
 
@@ -36,14 +43,14 @@ public enum AnnotationTaskType implements org.apache.thrift.TEnum {
    * Find a the enum type by its integer value, as defined in the Thrift IDL.
    * @return null if the value is not found.
    */
-  public static AnnotationTaskType findByValue(int value) { 
+  public static SummarySourceType findByValue(int value) { 
     switch (value) {
+      case 0:
+        return DOCUMENT;
       case 1:
-        return TRANSLATION;
+        return TOKENIZATION;
       case 2:
-        return NER;
-      case 3:
-        return TOPICID;
+        return ENTITY;
       default:
         return null;
     }